+++ CHANGES.rst
@@ -42,6 +42,7 @@ Improvements
   given editable type (:pr:`5755`)
 - Show only "ready to review" editables on the "get next editable" list (:pr:`5765`)
 - Disallow uploading empty files (:pr:`5767`)
+- Include non-speaker authors in the timetable export API (:issue:`5412`, :pr:`5738`)
 
 Bugfixes
 ^^^^^^^^